test on both s3 and fs (#21)

+++ test/init.js
@@ -1,5 +1,7 @@
 'use strict'
 
+process.env.NPM_REGISTER_FS_DIRECTORY = 'tmp'
 const fs = require('fs-extra')
-
-fs.outputFileSync('./tmp/htpasswd', 'test:$2y$05$ZhGKbrjyUbSbiMUeYeRUKOXPKzs9./NIZHsycrQkUKIj1Z2VybqdK\n')
+if (!fs.existsSync('./tmp/htpasswd')) {
+  fs.outputFileSync('./tmp/htpasswd', 'test:$2y$05$ZhGKbrjyUbSbiMUeYeRUKOXPKzs9./NIZHsycrQkUKIj1Z2VybqdK\n')
+}
